Refactor: Extract Report models to models module
Extracted `ReportStatus`, `ReportSection`, `ReportOutline`, `Report`,
`EvidenceItem`, and `ReportClaim` from the monolithic
`app/services/report_agent.py` into a new `app/models/report.py` file.

Updated the module `__init__.py` in models to export the new classes and
updated imports in `report_agent.py`, `report.py`, `runs.py`, and
affected test files.

All backend tests pass successfully after the change.
