+++ backend/app/models/__init__.py
@@ -4,6 +4,10 @@ Data Models Module
 
 from .task import TaskManager, TaskStatus
 from .project import Project, ProjectStatus, ProjectManager
+from .report import Report, ReportStatus, ReportOutline, ReportSection, EvidenceItem, ReportClaim
 
-__all__ = ['TaskManager', 'TaskStatus', 'Project', 'ProjectStatus', 'ProjectManager']
-
+__all__ = [
+    'TaskManager', 'TaskStatus',
+    'Project', 'ProjectStatus', 'ProjectManager',
+    'Report', 'ReportStatus', 'ReportOutline', 'ReportSection', 'EvidenceItem', 'ReportClaim'
+]
